requirements.txt
sqlalchemy
tornado
#fastapi
uvicorn
pymysql
bcrypt
cryptography
pyjwt
pandas
